Move updateBodyClass method
https://youtrack.weseek.co.jp/issue/GW-7844
- Move and rename updateBodyClass method
